insurance_agent/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
python-dotenv
langgraph
langchain-openai
langchain-core
twilio
sendgrid
requests
tenacity
elevenlabs
python-multipart
ecdsa
apscheduler
email-validator
httpx
openpyxl
